% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_freq}
\alias{calc_freq}
\title{Calculate frequencies of `col`-values.}
\usage{
calc_freq(data, col, omit_NA = TRUE)
}
\arguments{
\item{data}{A tibble}

\item{col}{Column name for which frequencies are calculated}

\item{omit_NA}{If TRUE, ignores NAs.}
}
\value{
Tibble with group names, their quantity (n) and frequencies (freq)
}
\description{
Calculate frequencies of `col`-values.
}
