% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calculate_UICC}
\alias{calculate_UICC}
\title{Calculate UICC stage}
\usage{
calculate_UICC(data, T, N, M)
}
\arguments{
\item{data}{A tibble}

\item{T}{Tumor size column name, values: "1", "2", "3", "4"}

\item{N}{Nodal status column name, values: "0", "1", "2", "3"}

\item{M}{Metastases column name, values: "M0", "M1"}
}
\value{
Same data with new UICC column containing UICC stage. If UICC cannot be calculated, it will be filled with NA
}
\description{
Calculate UICC stage
}
