% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{check_integrity}
\alias{check_integrity}
\title{Count missing values}
\usage{
check_integrity(data, T, N, M, D, U)
}
\arguments{
\item{data}{A tibble}

\item{T}{Tumor size column name}

\item{N}{Nodal status column name}

\item{M}{Metastases column name}

\item{D}{Grading column name}

\item{U}{UICC column name}
}
\value{
A tibble with three columns: full (which case of missing value), n (count), x (dummy for easy plotting)
}
\description{
Count missing values
}
