% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grading.R
\name{combine_diff}
\alias{combine_diff}
\title{combine 1/2 with 2, 2/3 with 3 and 3/4 with 4 for grading.}
\usage{
combine_diff(data, D)
}
\arguments{
\item{data}{A dataframe}

\item{D}{Column with grading}
}
\value{
Same dataframe with combined gradings.
}
\description{
combine 1/2 with 2, 2/3 with 3 and 3/4 with 4 for grading.
}
