% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{count_NAs}
\alias{count_NAs}
\title{Counts NAs.}
\usage{
count_NAs(data, T, N, M, D)
}
\arguments{
\item{data}{A tibble}

\item{T}{Tumor size column name}

\item{N}{Nodal status column name}

\item{M}{Metastases column name}

\item{D}{Grading column name}
}
\value{
A tibble with value, name, n columns containing number of NAs and whole data
}
\description{
Counts NAs.
}
