% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{create_age_groups}
\alias{create_age_groups}
\title{Create age groups}
\usage{
create_age_groups(data, age_name, res_name, ...)
}
\arguments{
\item{data}{A tibble with age column}

\item{age_name}{Name of the column with age data}

\item{res_name}{Name of the created column with age groups}

\item{...}{Age group borders in increasing order}
}
\value{
Same data with new res_name column containing age group info
}
\description{
Create age groups
}
