% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Donuts.R
\name{create_circle_coords}
\alias{create_circle_coords}
\title{Generate dim(data)[1] points around center with c0 coordinates and r radius.}
\usage{
create_circle_coords(data, c0 = c(0, 0), r = 1)
}
\arguments{
\item{data}{A dataframe}

\item{c0}{Center point of the generated circle}

\item{r}{Circle radius}
}
\value{
Same dataset with new x and y columns
}
\description{
Generate dim(data)[1] points around center with c0 coordinates and r radius.
}
