% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{create_coord_dataset}
\alias{create_coord_dataset}
\title{Create dataset for "tissue"-plotting}
\usage{
create_coord_dataset(data, thickness = 10, width = 4, height = 2.5)
}
\arguments{
\item{data}{A tibble with a `n` column at least}

\item{thickness}{Tissue's width given in cell count}

\item{width}{Cell width}

\item{height}{Cell height}
}
\value{
A tibble with coordinates of each cell within the tissue and corresponding value of variable of interest
}
\description{
Create dataset for "tissue"-plotting
}
