% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grading.R
\name{create_diff_dataset}
\alias{create_diff_dataset}
\title{Create dataset for grading's plots (only for 4 separate grading levels: 1, 2, 3 and 4)}
\usage{
create_diff_dataset(
  data,
  D,
  thickness = 10,
  width = 4,
  height = 2.5,
  base_noise = 0.55,
  noise_multipliers = c(0, 1, 2, 4)
)
}
\arguments{
\item{data}{A tibble with Grading and their count}

\item{D}{Column with grading}

\item{thickness}{Tissue's width given in cell count}

\item{width}{Cell width}

\item{height}{Cell height}

\item{base_noise}{Base noise for all grading steps}

\item{noise_multipliers}{Multipliers for every grading level}
}
\value{
A tibble with coordinates of each cell within the tissue, their grading status and cell nucleous size
}
\description{
Create dataset for grading's plots (only for 4 separate grading levels: 1, 2, 3 and 4)
}
