% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Donuts.R
\name{create_donut}
\alias{create_donut}
\title{Draw a donut graph with round edges}
\usage{
create_donut(data, col_group, col_feature)
}
\arguments{
\item{data}{A dataframe}

\item{col_group}{A column that divides data in **two** groups}

\item{col_feature}{A column containing information to be plotted}
}
\value{
A dataset appropriate for donut plotting
}
\description{
Draw a donut graph with round edges
}
