% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{new_size}
\alias{new_size}
\title{Change dataset size based on frequencies}
\usage{
new_size(data, total_num = 100, highlight_small = TRUE)
}
\arguments{
\item{data}{A tibble with a "freq"-column}

\item{total_num}{New size of the dataset}

\item{highlight_small}{If TRUE, adds +1 by rounding to the smallest frequencies at first; starts with the biggest by FALSE.}
}
\value{
same dataset with a new n column with fixed numbers
}
\description{
Change dataset size based on frequencies
}
